Finished

+++ workbook 2.11/src/main/java/com/ltp/workbook/Record.java
@@ -1,11 +1,13 @@
 package com.ltp.workbook;
 
+import java.math.BigDecimal;
+
 public class Record {
     private String item;
-    private double revenue;
-    private double cost;
+    private BigDecimal revenue;
+    private BigDecimal cost;
 
-    public Record(String item, double revenue, double cost) {
+    public Record(String item, BigDecimal revenue, BigDecimal cost) {
         this.item = item;
         this.revenue = revenue;
         this.cost = cost;
@@ -19,19 +21,19 @@ public class Record {
         this.item = item;
     }
 
-    public double getRevenue() {
+    public BigDecimal getRevenue() {
         return revenue;
     }
 
-    public void setRevenue(double revenue) {
+    public void setRevenue(BigDecimal revenue) {
         this.revenue = revenue;
     }
 
-    public double getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
 
-    public void setCost(double cost) {
+    public void setCost(BigDecimal cost) {
         this.cost = cost;
     }
 }
